Resolved Most Comments
