+++ api/app.mjs
@@ -212,11 +212,12 @@ app.post('/hypothesis', async (req, res) => {
         return res.status(400).json({ error: 'Invalid request!' });
     }
     const ids = body?.ids ?? [];
+    const filters = body?.filters ?? [];
+    console.log(filters);
     const virusFamilies = body?.virusFamilies ?? [];
-    const pageStart = body?.pageStart ?? 0;
-    const pageEnd = body?.pageEnd ?? undefined;
     const identifiers = body?.identifiers ?? [];
-    const result = await getMwasHypothesis(ids, virusFamilies, pageStart, pageEnd, identifiers);
+    const selectedMetadata = body?.selectedMetadata ?? [];
+    const result = await getMwasHypothesis(ids, virusFamilies, identifiers, filters, selectedMetadata);
     if (result.error) {
         console.error(result.error);
         return res.status(500).json({ error: result.error });
