+++ frontend/src/common/utils/textFormatting.ts
@@ -1,3 +1,6 @@
+import React from 'react';
+import Link from '@mui/material/Link';
+
 export const capitalize = (s: string) => {
     if (typeof s !== 'string') return '';
     return s.charAt(0).toUpperCase() + s.slice(1);
