+++ frontend/src/features/Figures/Virome/ViromeMWAS.tsx
@@ -84,7 +84,7 @@ const ViromeMWAS = ({ identifiers, virusFamilies }) => {
                 <Typography variant='h6' component={'div'} sx={{ flex: 1.2 }}>
                     {`Metadata-wide association study (MWAS) `}
                 </Typography>
-                <MwasHypothesisGenerator identifiers={identifiers} virusFamilies={virusFamilies}/>
+                <MwasHypothesisGenerator identifiers={identifiers} virusFamilies={virusFamilies} selectedMetadata={selectedMetadata}/>
                 {identifiers && identifiers['bioproject'].single.length > 100 ? (
                     <Typography variant='h7' component={'div'} sx={{ mt: 2, flex: 1 }}>
                         {`Dataset is too large. Displaying partial results.`}
