+++ frontend/src/features/LLM/MwasHypothesisGenerator.tsx
@@ -1,17 +1,21 @@
 import React from 'react';
+import { useSelector } from 'react-redux';
+
+import { selectAllFilters } from '../Query/slice.ts';
+import { getFilterQuery } from '../../common/utils/queryHelpers.ts';
 
 import { useLazyGetHypothesisQuery } from '../../api/client.ts';
 import { load } from '@fingerprintjs/botd';
 
-import IconButton from '@mui/material/IconButton';
-import BlurOnIcon from '@mui/icons-material/BlurOn';
 import Box from '@mui/material/Box';
 import Skeleton from '@mui/material/Skeleton';
 import Typography from '@mui/material/Typography';
-import Link from '@mui/material/Link';
-import Tooltip from '@mui/material/Tooltip';
+import { processGeneratedText } from './textFormatting.tsx';
+import { GenerateButton } from './GenerateButton.tsx';
 
-const MwasHypothesisGenerator = ({ identifiers, virusFamilies }) => {
+const MwasHypothesisGenerator = ({ identifiers, virusFamilies, selectedMetadata }) => {
+    const filters = useSelector(selectAllFilters);
+
     const [getHypothesisText, { data: hypothesisData, isLoading: isLoadingHypothesis, error: errorHypothesis }] =
         useLazyGetHypothesisQuery();
 
@@ -27,15 +31,15 @@ const MwasHypothesisGenerator = ({ identifiers, virusFamilies }) => {
                 ids: identifiers ? identifiers['bioproject'].single : [],
                 idRanges: identifiers ? identifiers['bioproject'].range : [],
                 virusFamilies: virusFamilies ? virusFamilies : [],
-                pageStart: 0,
-                pageEnd: 1000,
                 identifiers: identifiers,
+                filters: getFilterQuery({ filters }),
+                selectedMetadata: selectedMetadata,
             });
         } catch (error) {
             console.error(error);
         }
     };
-    console.log(hypothesisData);
+
     const renderPlaceholder = () => {
         if (isLoadingHypothesis) {
             return <Skeleton variant='text' width={'100%'} height={60} />;
@@ -43,44 +47,6 @@ const MwasHypothesisGenerator = ({ identifiers, virusFamilies }) => {
         return null;
     };
     const hypothesisTextIsNonEmpty = () => hypothesisData && hypothesisData?.text?.length > 0;
-
-    const processHypothesisText = (text) => {
-        const combinedRegex = /PRJ\w+|\|([^|]+)\|/g;
-        const combinedMatches = text.match(combinedRegex);
-        if (combinedMatches) {
-            const finalComponents: (string | JSX.Element)[] = [];
-            let lastIndex = 0;
-            combinedMatches.forEach((match, index) => {
-                const matchIndex = text.indexOf(match, lastIndex);
-                if (matchIndex > lastIndex) {
-                    finalComponents.push(text.substring(lastIndex, matchIndex));
-                }
-                if (/PRJ\w+/.test(match)) {
-                    finalComponents.push(
-                        <Link
-                            key={`link-${index}`}
-                            href={`https://www.ncbi.nlm.nih.gov/bioproject/${match}`}
-                            target='_blank'
-                        >
-                            {match}
-                        </Link>,
-                    );
-                } else if (/\|([^|]+)\|/.test(match)) {
-                    finalComponents.push(
-                        <span key={`bold-${index}`} style={{ fontWeight: 'bold' }}>
-                            {match.slice(1, -1)}
-                        </span>,
-                    );
-                }
-                lastIndex = matchIndex + match.length;
-            });
-            if (lastIndex < text.length) {
-                finalComponents.push(text.substring(lastIndex));
-            }
-            return finalComponents;
-        }
-        return text;
-    };
 
     return (
         <Box>
@@ -94,16 +60,7 @@ const MwasHypothesisGenerator = ({ identifiers, virusFamilies }) => {
                     mr: hypothesisTextIsNonEmpty() ? 0 : '2%',
                 }}
             >
-                <Tooltip title='Generate MWAS Hypotheses' placement='bottom'>
-                    <IconButton style={{ backgroundColor: 'rgba(86, 86, 86, 0.4)' }} onClick={onButtonClick}>
-                        <BlurOnIcon
-                            style={{
-                                color: '#9be3ef',
-                                fontSize: 30,
-                            }}
-                        />
-                    </IconButton>
-                </Tooltip>
+            {GenerateButton({ onButtonClick, title: 'Generate Hypothesis' })}
             </Box>
             <Box
                 sx={{
@@ -117,7 +74,7 @@ const MwasHypothesisGenerator = ({ identifiers, virusFamilies }) => {
                 {renderPlaceholder()}
                 {hypothesisTextIsNonEmpty() ? (
                     <Typography variant='body' sx={{ mt: 2, mb: 4, whiteSpace: 'pre-wrap' }}>
-                        {processHypothesisText(hypothesisData.text)}
+                        {processGeneratedText(hypothesisData.text)}
                     </Typography>
                 ) : null}
             </Box>
